Improve the functionality of the schema resolver feature
Fix #216

Refactored things so that the ResourceRequest and ResourceResponse are
interfaces. Reworked the ExampleSchemaResolver to create a custom
ResourceResponse to demonstrate resolving a customer URI scheme.
