+++ gradle.properties
@@ -2,7 +2,7 @@ basename=xmlresolver
 
 # ************************************************
 # When this version number changes:
-resolverVersion=6.0.11
+resolverVersion=6.0.12-SNAPSHOT
 # Also change the version number in overview.html
 # FIXME: figure out how to automate this.
 # ************************************************
