+++ src/main/java/org/xmlresolver/adapters/SAX1Adapter.java
@@ -4,10 +4,7 @@ import org.xml.sax.EntityResolver;
 import org.xml.sax.InputSource;
 import org.xml.sax.SAXException;
 import org.xml.sax.ext.EntityResolver2;
-import org.xmlresolver.ResolverConstants;
-import org.xmlresolver.ResourceRequest;
-import org.xmlresolver.ResourceResponse;
-import org.xmlresolver.XMLResolver;
+import org.xmlresolver.*;
 import org.xmlresolver.sources.ResolverInputSource;
 
 import java.io.IOException;
