+++ src/main/java/org/xmlresolver/loaders/ValidatingXmlLoader.java
@@ -71,7 +71,7 @@ public class ValidatingXmlLoader implements CatalogLoader {
                 return catalogMap.get(catalog);
             }
             try {
-                ResourceRequest request = new ResourceRequest(config);
+                ResourceRequest request = new ResourceRequestImpl(config);
                 request.setURI(catalog);
                 request.setOpenStream(true);
                 ResourceResponse resp = config.getResource(request);
