+++ src/main/java/org/xmlresolver/sources/ResolverLSInput.java
@@ -8,7 +8,6 @@ import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.Reader;
 import java.net.URI;
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
